requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
configparser==5.0.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
hs-infra==0.0.9
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
minio==6.0.0
Pillow==8.0.1
pkg-resources==0.0.0
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.1
wrapt==1.12.1
